requirements.txt
pandas==1.1.0
requests
beautifulsoup4
datetime
joblib
urllib3
sqlalchemy
xlsxwriter
openpyxl
holidays
psycopg2-binary
lxml
xlrd
openpyxl
tqdm
python-dotenv
pythonnet
xlsxwriter
xlrd
openpyxl
holidays
scipy
